// src/interfaces/salesRepId.interface.ts
export interface ISalesRep {
  S_Number: number;
  S_Desc?: string | null;

  Commission_Rate01?: number | null;
  Commission_Rate02?: number | null;
  Commission_Rate03?: number | null;
  Commission_Rate04?: number | null;
  Commission_Rate05?: number | null;
  Commission_Rate06?: number | null;
  Commission_Rate07?: number | null;
  Commission_Rate08?: number | null;
  Commission_Rate09?: number | null;
  Commission_Rate10?: number | null;
  Commission_Rate11?: number | null;
  Commission_Rate12?: number | null;

  Commission_Type01?: string | null;
  Commission_Type02?: string | null;
  Commission_Type03?: string | null;
  Commission_Type04?: string | null;
  Commission_Type05?: string | null;
  Commission_Type06?: string | null;
  Commission_Type07?: string | null;
  Commission_Type08?: string | null;
  Commission_Type09?: string | null;
  Commission_Type10?: string | null;
  Commission_Type11?: string | null;
  Commission_Type12?: string | null;
}